requirements.txt
asgiref==3.3.4
beautifulsoup4==4.9.3
Django==3.2.3
django-bootstrap4==3.0.1
django-crispy-forms==1.12.0
django-durationwidget2==1.0.6
django-filter==2.4.0
gunicorn==20.1.0
pytz==2021.1
soupsieve==2.2.1
sqlparse==0.4.1
